+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - teste de funcionamento da função reduce
 - exercício erros
 - montagem do ambiente para o exercício consumindo APIs
+- script.js para a captura da imagem por metodo fetch em uma url.
 
 ## 1.1.0 / 15/06/2022
 - resolução do desafio de comparação usando operadores e sintaxe